backend/requirements.txt
Flask>=2.0.0
flask-cors>=3.0.0
google-cloud-vision>=3.0.0
Pillow>=9.0.0
pytesseract>=0.3.8
numpy>=1.21.0
groq>=0.4.0
python-dotenv>=1.0.0
gunicorn>=20.0.0
boto3>=1.28.0
botocore>=1.31.0
